fix: fix country map (#46)
