+++ frontend/package.json
@@ -17,6 +17,7 @@
     "@typescript-eslint/parser": "4.8.1",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.4.1",
     "axios": "^0.21.1",
+    "country-iso-2-to-3": "^1.1.0",
     "enzyme": "3.11.0",
     "eslint": "7.11.0",
     "eslint-config-airbnb-typescript": "12.0.0",
@@ -26,8 +27,8 @@
     "eslint-plugin-prettier": "3.1.4",
     "eslint-plugin-react": "7.21.5",
     "eslint-plugin-react-hooks": "4.0.8",
-    "leaflet": "^1.7.1",
     "iso-country-currency": "^0.5.3",
+    "leaflet": "^1.7.1",
     "prettier": "2.0.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
