+++ frontend/src/components/CountryPage/CountryPage.tsx
@@ -15,7 +15,7 @@ import { homeStyles } from '../Home/HomeStyles';
 import { setActiveCountryData, setCountryId, toogleIsLoading } from '../../redux/countriesReducer';
 import Weather from '../Weather/Weather';
 import Preloader from '../Preloader/Preloader';
-// import { Map } from '../CountryMap/Map';
+import Map from '../Map/Map';
 import Video from '../Video/Video';
 
 const CountryPage = ({
@@ -96,9 +96,9 @@ const CountryPage = ({
             <ExchangeRates />
           </div>
           <Video src={video} />
+          <Map map={map} />
         </>
       )}
-      {/* <Map map={map} /> */}
     </div>
   );
 };
